Functional os ver0.1
