+++ dpdd.ino
@@ -1,4 +1,5 @@
 #include "hardware.h"
+#include "display.h"
 #include "game.h"
 #include "os.h"
 
@@ -6,8 +7,9 @@ OS os;
 
 void setup() {
   init_hardware();
+  Display::updateDisplay(&os);
 }
 
 void loop() {
-  os.loop();
+  os.loop();  
 }
